+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Implement `TryFrom<&str>` for `{Array,Group}Metadata`
 
 ### Changed
- - Make `array::codec::array_to_bytes::bytes::Endianness::{is_native(),NATIVE_ENDIAN}` public
+ - Make `array::codec::array_to_bytes::bytes::{Endianness::is_native(),NATIVE_ENDIAN,reverse_endianness}` public
 
 ## [0.12.4] - 2024-03-09